requirements.txt
numpy
opencv-python
psutil
absl-py
pyenet
scipy
pygame
pandas
easydict
pyenet
libconf